Store user details on the profile, to make checking for an email initially easier
